+++ tests/_test_process_executor.py
@@ -116,7 +116,7 @@ class ExecutorShutdownTest:
             e.submit(id, ExitAtPickle())
 
     def test_interpreter_shutdown(self):
-        # Free ressources to avoid random timeout in CI
+        # Free resources to avoid random timeout in CI
         self.executor.shutdown(wait=True, kill_workers=True)
 
         tempdir = tempfile.mkdtemp(prefix='loky_')
@@ -892,3 +892,23 @@ class ExecutorTest:
         # Submitting other jobs fails as well.
         with pytest.raises(TerminatedWorkerError, match=match):
             self.executor.submit(pow, 2, 8)
+
+    @staticmethod
+    def _test_child_env(var_name):
+        import os
+        return os.environ.get(var_name, "unset")
+
+    def test_child_env_executor(self):
+        # Test that for loky context, setting argument env correctly overwrite
+        # the environment of the child process.
+        if self.context.get_start_method() != 'loky':
+            pytest.skip(msg="Only work with loky context")
+
+        var_name = "loky_child_env_executor"
+        var_value = "variable set"
+        executor = self.executor_type(1, env={var_name: var_value})
+
+        var_child = executor.submit(self._test_child_env, var_name).result()
+        assert var_child == var_value
+
+        executor.shutdown(wait=True)