+++ tests/test_loky_backend.py
@@ -5,6 +5,7 @@ import psutil
 import pytest
 import signal
 import pickle
+import platform
 import socket
 import multiprocessing as mp
 from tempfile import mkstemp
@@ -275,6 +276,46 @@ class TestLokyBackend:
         parent_connection.close()
         child_connection.close()
 
+    @staticmethod
+    def _test_child_env(key, queue):
+        import os
+
+        queue.put(os.environ.get(key, 'not set'))
+
+    @pytest.mark.xfail(sys.version_info[0] == 2 and
+                       sys.platform == "win32",
+                       reason="Can randomly fail on python2.7 and windows.")
+    def test_child_env_process(self):
+        import os
+
+        key = 'loky_child_env_process'
+        value = 'loky works'
+        out_queue = self.SimpleQueue()
+        try:
+            # Test that the environment variable is correctly copied in the
+            # child process.
+            os.environ[key] = value
+            p = self.Process(target=self._test_child_env,
+                             args=(key, out_queue))
+            p.start()
+            child_var = out_queue.get()
+            p.join()
+
+            assert child_var == value
+
+            # Test that the environment variable is correctly overwritted by
+            # using the `env` argument in Process.
+            new_value = 'loky rocks'
+            p = self.Process(target=self._test_child_env,
+                             args=(key, out_queue), env={key: new_value})
+            p.start()
+            child_var = out_queue.get()
+            p.join()
+
+            assert child_var == new_value, p.env
+        finally:
+            del os.environ[key]
+
     @classmethod
     def _test_terminate(cls, event):
         # Notify the main process that child process started
@@ -511,6 +552,11 @@ class TestLokyBackend:
 
         return named_sem
 
+    @pytest.mark.skipif(
+        platform.python_implementation() == "PyPy" and
+        sys.version_info[:3] <= (3, 5, 3),
+        reason="early PyPy versions leak a file descriptor, see "
+               "https://bitbucket.org/pypy/pypy/issues/3021")
     def test_sync_object_handling(self):
         """Check the correct handling of semaphores and pipes with loky