+++ CHANGES.md
@@ -6,8 +6,8 @@
 - Copy the environment variables in the child process for ``LokyProcess``. Also
   add a ``env`` argument in ``LokyProcess``, ``ProcessPoolExecutor`` and
   ``get_reusable_executor`` to over-write consistently some environment variable
-  in the child process. This allow setting env variables before loading any
-  module. This feature is unstable with Python2.7 on windows. (#217)
+  in the child process. This allows setting env variables before loading any
+  module. This feature is unreliable on Windows with Python 2.7. (#217)
 
 
 ### 2.5.1 - 2019-06-11 - Bugfix release